Handling for global | local instalation integrated
This function checks if a package "esdoc-standard-plugin" is installed in parallel with this instance of @enterthenamehere/esdoc.
Normally this is so, if the actual node package instance of @enterthenamehere/esdoc is a global or global-style instance (installed with 'npm i -g ...').
It is necessary to use the esdoc-core of this parallel instance (esdoc-standard-plugin) so that the plugins of esdoc-standard-plugin can use the same esdoc-core and thus works globaly in every project (without localy installing esdoc in the projects).
So the Developer can use esdoc inside a project or as global installed node package ('npm i -g ...').
Normally all global node packages will be installed in one node-folder (e.g. node_modules) in a os specific path (e.g. in Linux: '/usr/local/lib/node_modules').