Fix whitespace-bug with tag values
Before this update, it was not possible, for example, to use "@access" in the tag name with a tab feed or more than one space between the tag name and tag value. If you used more than one space, this part (e.g. a variable) was not documented at all.